Implement Static Hoisting
The `hoist` option hoists static attributes to the `Program` node to
minimize allocations. When a non-static key is also present, we will
assign the key to the hoisted array in the correct value position.

We purposefully deopt when we need to wrap the JSX Element, since
deferred execution will screw things up.